MAGO-56 added util function for drawing boxed text

+++ src/app/applayer.h
@@ -10,8 +10,8 @@ public:
 	AppLayer();
 	~AppLayer() override;
 
-	void on_update(const int dt) override;
-	void on_render(const int dt) override;
+	void on_update(const float dt) override;
+	void on_render(const float dt) override;
 
 private:
 	std::unique_ptr<AppLayerPriv> _p;
